[DEVOPS] chore: alterando tratamento imagens

+++ tratamentoImagem/src/app.py
@@ -2,9 +2,9 @@ import boto3
 import cv2
 import os
 import numpy as np
+from PIL import Image
 from botocore.exceptions import NoCredentialsError
-
-# boto3.setup_default_session(profile_name="faculdade")
+from sklearn.decomposition import PCA
 
 # Função para baixar uma imagem do S3
 def download_image(bucket_name, image_key, local_filename):
@@ -13,12 +13,34 @@ def download_image(bucket_name, image_key, local_filename):
         s3.download_file(bucket_name, image_key, local_filename)
     except NoCredentialsError:
         print("Credenciais não configuradas corretamente ou não disponíveis")
+        return False
+    return True
 
 # Função para carregar uma imagem em tons de cinza
 def load_image_grayscale(filename):
-    image = cv2.imread(filename)
-    grayscale_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-    return grayscale_image
+    try:
+        image = Image.open(filename).convert('L')
+        img_array = np.array(image)
+        return img_array
+    except Exception as e:
+        print(f"Erro ao carregar a imagem {filename}: {e}")
+        return None
+
+# Função para aplicar PCA
+def apply_pca(image, n_components=30):
+    original_img = Image.open(image).convert('L')
+    img_array = np.array(original_img)
+
+    original_components = img_array.shape[1]
+
+    n_components = n_components
+
+    pca = PCA(n_components=n_components)
+    pca.fit(img_array)
+    img_transformed = pca.transform(img_array)
+    img_reconstructed = pca.inverse_transform(img_transformed)
+    processed_components = pca.n_components_
+    return img_reconstructed
 
 # Função para fazer upload de uma imagem para o S3
 def upload_image(bucket_name, image_key, local_filename):
@@ -28,21 +50,8 @@ def upload_image(bucket_name, image_key, local_filename):
     except NoCredentialsError:
         print("Credenciais não configuradas corretamente ou não disponíveis")
 
-# Função para solicitar o caminho da pasta ao usuário
-def get_folder_path():
-    return input("Por favor, insira o caminho da pasta: ")
-
-# Solicite o caminho da pasta ao usuário
-folder_path = "Acne"
-
-# Verifique se o caminho da pasta é válido
-# if not os.path.isdir(folder_path):
-#     print(f"O caminho {folder_path} não é uma pasta válida.")
-#     exit(1)
-
 # Defina os nomes dos buckets da AWS
 bucket_raw = 'tcc-dev-raw-bucket'
-bucket_consumed = 'tcc-dev-consumed-bucket'
 bucket_staged = 'tcc-dev-staged-bucket'
 
 # Conecte-se ao serviço S3
@@ -52,21 +61,30 @@ s3 = boto3.resource('s3')
 raw_bucket = s3.Bucket(bucket_raw)
 raw_images = [obj.key for obj in raw_bucket.objects.all()]
 
-# Obtenha a lista de chaves de objetos no bucket consumed
+# Obtenha a lista de chaves de objetos no bucket staged
 staged_bucket = s3.Bucket(bucket_staged)
 staged_images = [obj.key for obj in staged_bucket.objects.all()]
 
 # Verifique se há novas imagens no bucket raw
 for image_key in raw_images:
     if image_key not in staged_images:
-        local_filename = os.path.join('temp_image.jpg')
+        local_filename = 'temp_image.jpg'
         # Baixe a imagem do bucket raw
-        download_image(bucket_raw, image_key, local_filename)
-        # Carregue a imagem em tons de cinza
-        grayscale_image = load_image_grayscale(local_filename)
-        # Salve a imagem processada no bucket staged
-        processed_image_key = 'Acne/' + image_key  # Adicione um prefixo para identificar imagens processadas
-        cv2.imwrite(local_filename, grayscale_image)  # Sobrescreve a imagem original com a versão em tons de cinza
-        upload_image(bucket_staged, image_key, local_filename)
-        print(f"Imagem {image_key} processada e salva no bucket staged")
-        os.remove(local_filename)
+        if download_image(bucket_raw, image_key, local_filename):
+            # Verifique se a imagem foi baixada corretamente
+            if os.path.getsize(local_filename) > 0:
+                # Carregue a imagem em tons de cinza
+                grayscale_image = load_image_grayscale(local_filename)
+                if grayscale_image is not None:
+                    # Aplique PCA
+                    pca_image = apply_pca(grayscale_image)
+                    # Converta para uint8
+                    pca_image_uint8 = pca_image.astype(np.uint8)
+                    # Salve a imagem processada no bucket staged
+                    processed_image_key = image_key  # Preserve a estrutura do caminho
+                    Image.fromarray(pca_image_uint8).save(local_filename)  # Salva a imagem PCA
+                    upload_image(bucket_staged, processed_image_key, local_filename)
+                    print(f"Imagem {image_key} processada e salva no bucket staged")
+                os.remove(local_filename)
+            else:
+                print(f"Falha ao baixar a imagem {image_key}")
